fixed a "comparison always ..."
git-svn-id: https://svn.aros.org/svn/aros/trunk@8531 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/aminet/games/think/lazymines/lazymines.c
@@ -153,7 +153,7 @@ main (
    pubscr_name[0] = '\0';
    
    handle_startup_msg (argv, argc == 0);
-   if (auto_opening < 0)
+   if ((BYTE)auto_opening < 0)
       auto_opening = 0;
    else if (auto_opening > 10)
       auto_opening = 10;
